Refactor code to be more OOP with new UI, Storage, Parser and Tasklist classes
