+++ data/bill.txt
@@ -1,2 +1,5 @@
-1.[T][ ] readbook
-2.[D][ ] run marathon (by: Mar 10 2012)
+1.[T][X] crying
+2.[D][ ] cryhard (by: Oct 08 2024)
+3.[D][X] jump (by: Dec 10 2024)
+4.[T][X] fly
+5.[T][ ] sad
